clean & reorganize code
