+++ seismicpro/src/const.py
@@ -20,7 +20,6 @@ ENDIANNESS = {
 
 # Custom headers optionally added by SeismicPro to those from SEG-Y file
 HDR_DEAD_TRACE = 'DeadTrace'
-HDR_CLIP = 'Clips'
 HDR_FIRST_BREAK = 'FirstBreak'
