+++ seismicpro/src/utils/general_utils.py
@@ -5,7 +5,6 @@ from concurrent.futures import Future, Executor
 import numpy as np
 
 from numba import njit
-from scipy import signal, fft
 
 import numpy as np
 
@@ -183,133 +182,6 @@ def dump_header(obj_with_headers, path, header_col, trace_id_cols=('FieldRecord'
         f.write(rows_as_str)
 
 
-
-@njit(nogil=True)
-def strip_clip_indicator(clip_ind):
-    n = len(clip_ind)
-    for i0 in range(n):
-        if not clip_ind[i0]:
-            break
-
-    i1 = n
-    for j in range(n - 1 - i0):
-        if not clip_ind[n - 1 - j]:
-            break
-        else:
-            i1 -= 1
-    return i0, i1
-
-
-@njit(nogil=True)
-def get_cliplen_indicator(traces):
-    traces = np.atleast_1d(traces)
-
-    indicator = (traces[..., 1:] == traces[..., :-1]).astype(np.int32)
-
-    for i in range(1, indicator.shape[-1]):
-        indicator[..., i] += indicator[..., i-1] * (indicator[..., i] == 1)
-
-    return indicator
-
-
-@njit(nogil=True)
-def get_clip_indicator(traces, clip_len):
-    traces = np.atleast_1d(traces)
-
-    n_samples = traces.shape[0]
-
-    if clip_len < 2 or clip_len > n_samples - 1:
-        raise ValueError("Incorrect `clip_len`")
-
-    ind_len = n_samples - clip_len + 1
-
-    clip_indicator = np.full_like(traces[:ind_len], True, dtype=np.bool8)
-    for curr_shift in range(1, clip_len):
-        clip_indicator &= (traces[curr_shift:ind_len + curr_shift] == traces[:ind_len])
-
-    return clip_indicator
-
-
-@njit(nogil=True)
-def has_clips(trace, clip_len):
-
-    trace = np.asarray(trace)
-    if trace.ndim != 1:
-        raise ValueError("Only 1-D traces are allowed")
-
-    clip_res = get_clip_indicator(trace, clip_len)
-
-    i0, i1 = strip_clip_indicator(clip_res)
-
-    return np.any(clip_res[i0:i1])
-
-
-def get_maxabs_clips(traces):
-    maxes = traces.max(axis=-1, keepdims=True)
-    mins = traces.min(axis=-1, keepdims=True)
-
-    res_plus = np.isclose(traces, maxes, atol=0)
-    res_minus = np.isclose(traces, mins, atol=0)
-
-    return ((res_plus[..., :-2] & res_plus[..., 1:-1] & res_plus[..., 2:])
-            | (res_minus[..., :-2] & res_minus[..., 1:-1] & res_minus[..., 2:]))
-
-def has_maxabs_clips(traces):
-    return np.any(get_maxabs_clips(traces), axis=-1)
-
-def calc_spikes(arr):
-    with fft.set_workers(25):
-        running_mean = signal.fftconvolve(arr, [[1,1,1]], mode='valid', axes=1)/3
-    return (np.abs(arr[...,1:-1] - running_mean))
-
-@njit
-def fill_nulls(arr):
-
-    n_samples = arr.shape[1]
-
-    for i in range(arr.shape[0]):
-        nan_indices = np.nonzero(np.isnan(arr[i]))[0]
-        if len(nan_indices) > 0:
-            j = nan_indices[-1]+1
-            if j < n_samples:
-                arr[i, :j] = arr[i, j]
-
-@njit(nogil=True)
-def get_const_indicator(traces, cmpval=None):
-
-    if cmpval is None:
-        indicator = (traces[..., 1:] == traces[..., :-1])
-    else:
-        indicator = (traces[..., 1:] == cmpval)
-
-    brdr_zeros = np.zeros(traces.shape[:-1]+(1,), dtype=np.bool8)
-    indicator = np.concatenate((brdr_zeros, indicator), axis=-1)
-
-    return indicator.astype(np.int32)
-
-@njit(nogil=True)
-def get_constlen_indicator(traces, cmpval=None):
-
-    old_shape = traces.shape
-
-    traces = np.atleast_2d(traces)
-
-    indicator = get_const_indicator(traces, cmpval)
-
-
-    for i in range(1, indicator.shape[-1]):
-        indicator[..., i] += indicator[..., i-1] * (indicator[..., i] == 1)
-
-
-    for j in range(0, indicator.shape[-2]):
-        for i in range(indicator.shape[-1] - 1, 0, -1):
-            if indicator[..., j, i] != 0 and indicator[..., j, i - 1] != 0:
-                indicator[..., j, i - 1] = indicator[..., j, i]
-
-    if cmpval is None:
-        indicator += (indicator > 0).astype(np.int32)
-
-    return indicator.reshape(*old_shape)
 
 class Coordinates:
     """Define spatial coordinates of an object."""
